Some bugs fixed. Still, matrixop will not compile in gcc and in icpc due to matrixop::minor issue with variadic macros.
git-svn-id: file:///mnt/svn/rtmath@74 1d52797c-35fc-48e8-8e80-2db72aa187e2

+++ rtmath/rtmath-base/quadrature.cpp
@@ -61,7 +61,7 @@ namespace rtmath {
 			// For each n degree, expect n quadrature points
 			// Start point in array is 2 + 3 + 4 + ... + n-1
 			// So, the starting index is 3 * 
-			unsigned int start = 3 * (-1 - degree / 2 + degree * degree/2);
+			unsigned int start = 3 * (unsigned int) (-1.0 - (double) degree / 2.0 + (double) degree * (double) degree / 2.0);
 			// Check if the start even exists! If not, generate the necessary points.
 			// TODO: add code. For now, throw if insufficient points
 			if (degree > 7) throw;
